Big peice got smaller chumk
